Implements min (water) and max (air) values for a sensor.
Add two variables for rolling min/max of the sensor reading. The max value (airMax) indicates no moisture and the min value (waterMin) indicates the sensor is submerged in water. Start these variables at a center point and they will make their way to the extremes as the sensor is submerged or dried. Keep these values to use as the calibration baseline for other projects using the sensors.